+++ examples/evals/fixture/src/handler.go
@@ -1,0 +1,3 @@
+package src
+
+func Handle() error { return nil }